pkg/capabilities move to daemon/internal
This package was added in 8f936ae8cf6c39bf3ce21a25b231383dac3212e6, and
never had external consumers. Let's move it internal.
