+++ daemon/devices.go
@@ -2,7 +2,7 @@ package daemon // import "github.com/docker/docker/daemon"
 
 import (
 	"github.com/docker/docker/api/types/container"
-	"github.com/docker/docker/pkg/capabilities"
+	"github.com/docker/docker/daemon/internal/capabilities"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 )
